Add pure json endpoint for related content
Also, move out shared code into models package.
